refactor: Split flyway scripts between common and tenant dependant
Scripts have been split this way due to the interleaved DDL and DML.

HEE is used by default and overridden in tests.
A number of the files in the NIMDTA directory are not strictly needed.
They are there for a check of flyway history and validation of changes.

TISNEW-3822: Migrate NIMDTA Data