test: Fix test issues related to uuid field
The uuid column is not created in the H2 database due to a dummy script
being used, add the column in the dummy script but exclude the triggers
and value population as they will not work in H2.
Update the test configuraton to skip schema validation in the H2
database, there is inconsistency in how the UUID is handled in H2 that
leads to validation failures.

TISNEW-5757