fix: make leaving reasons queryable
The endpoint responsible for answering with leaving reasons filtered
by parameters always responds with a full unfiltered list of reasons.
The findAll method needed a parameter for searchString.
There is now a test in place for querying leaving reasons by string.

TIS21-1582: Search Bar in the Trust Search Page not working