fix: create ObjectMapper from builder
ObjectMapper was created as a new instance and it's no longer
configured correctly (since spring-boot 2.5 ?). In order to have the
JavaTimeModule already configured it needs to be created via
Jackson2ObjectMapperBuilder. Since TIS-REFERENCE-service already has
the Audit module injected (where a util class for ObjectMappers is
present), it now creates an ObjectMapper via a that util class (which
uses a builder rather than creating a new instance)

TIS21-989: Trust linked to a Site not updating the TrustID linked to it when adding/updating a Site