feat: Site controller to deal with Current Trusts
The createSite() and updateSite() endpoints have a functionality to
fetch the trustId based on the trustCode.
In some cases there are several Trusts with the same trustCode, only
one of which is CURRENT. This fix will only retrieve the trustId of
CURRENT Trusts.

TIS21-1841: Create/Update Sites with CURRENT Trusts only