fix: Remove unused and vulnerable config/resource
The log management endpoints pose a security risk if the logging level
is changed by a bad actor, these endpoints are not currently used in any
legitimate way so should be removed.

LoggingConfiguration sets up logstash when enabled in the jhipster
configuration, however that is never the case for any deployed instance.
The configuration should be removed.

Add a TODO in AuditingAspect to review the use of reflection, it has
been flagged as a potential issue. While it is not believed to cause any
vulnerability, alternative techniques should be considered when
reviewing our overall auditing processes.

TIS21-2032